refactor(pagination): move PaginationList to typedef, add PageFetchResult alias, flatten tests
- Move PaginationList from pyiceberg/utils/pagination.py into pyiceberg/typedef.py
  alongside the project's other core types; delete the now-empty module
- Introduce _PageFetchResult TypeAlias for tuple[list[Identifier], str | None]
  in rest/__init__.py to avoid repeating the type on every _fetch_page signature
- Flatten TestPaginationList* classes into top-level functions to match the
  style used throughout the rest of the test suite
- Revert D105/D418 additions from .pre-commit-config.yaml; suppress D105
  inline on the two @overload stubs that cannot have docstrings
